修改删除方式和文件大小字段的类型

+++ src/main/java/com/devteam/mikufunbackend/entity/DandanPlayMatchRequestV0.java
@@ -12,7 +12,7 @@ import lombok.Data;
 public class DandanPlayMatchRequestV0 {
     String fileName;
     String fileHash;
-    int fileSize;
+    long fileSize;
     int videoDuration;
     // 'hashAndFileName', 'fileNameOnly', 'hashOnly'
     String matchMode;
